events permissions partially working

+++ events.php
@@ -72,8 +72,10 @@ $total_pages = ceil($total_events / $events_per_page);
                         <strong>Location:</strong> <?php echo htmlspecialchars($row['location']); ?>
                     </p>
                     <a href="details_event.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Details</a>
-                    <a href="edit_event.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Edit</a>
-                    <a href="delete_event.php?id=<?php echo $row['id']; ?>" class="btn btndanger" onclick="return confirm('Are you sure you want to delete this event?')">Delete</a>
+                    <?php if (isset($_SESSION['status']) && ($_SESSION['status'] === 'admin' || $_SESSION['status'] === 'mentor')): ?>
+                        <a href="edit_event.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Edit</a>
+                        <a href="delete_event.php?id=<?php echo $row['id']; ?>" class="btn btndanger" onclick="return confirm('Are you sure you want to delete this event?')">Delete</a>
+                    <?php endif; ?>
                 </div>
             </div>
         </div>
