+++ jobs.php
@@ -141,8 +141,10 @@ $total_pages = ceil($total_jobs / $jobs_per_page);
                         <strong>Type:</strong> <?php echo htmlspecialchars($row['type']); ?>
                     </p>
                     <a href="details_job.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Details</a>
-                    <a href="edit_job.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Edit</a>
-                    <a href="delete_job.php?id=<?php echo $row['id']; ?>" class="btn btndanger" onclick="return confirm('Are you sure?')">Delete</a>
+                    <?php if (isset($_SESSION['status']) && ($_SESSION['status'] === 'admin' || $_SESSION['status'] === 'mentor')): ?>
+                        <a href="edit_job.php?id=<?php echo $row['id']; ?>" class="btn btnprimary">Edit</a>
+                        <a href="delete_job.php?id=<?php echo $row['id']; ?>" class="btn btndanger" onclick="return confirm('Are you sure?')">Delete</a>
+                    <?php endif; ?>
                 </div>
             </div>
         </div>
